+++ src/blocking_ring_buffer.h
@@ -9,6 +9,7 @@
 #define SRC_BLOCKING_RING_BUFFER_H_
 
 #include "ring_buffer.h"
+#include <stdint.h>
 #include <pthread.h>
 
 class BlockingRingBuffer {
@@ -16,8 +17,8 @@ public:
 	BlockingRingBuffer(size_t);
 	virtual ~BlockingRingBuffer();
 
-	size_t ReadFrom(float *, size_t);
-	size_t WriteInto(float *, size_t);
+	size_t ReadFrom(uint8_t *, size_t);
+	size_t WriteInto(uint8_t *, size_t);
 
 	size_t FreeSpace();
 	size_t DataStored();
@@ -26,7 +27,7 @@ public:
 	void set_last_frame(bool);
 
 private:
-	RingBuffer<float> buffer_;
+	RingBuffer<uint8_t> buffer_;
 
 	pthread_mutex_t count_mutex_;
 	pthread_cond_t count_condition_not_empty_;
